Use promotion values provided by ArgoCD in post-sync workflow
This removes the check against the git repo for whether a given app should be promoted or not and replaces it with a check against values read from the Git repo at the time of the sync event. This fixes a race condition where a deployment could be unintentionally promoted to production.