Remove Contacts Admin charts
The app is being retired.

Note that the Google Tag Manager bootstrap GTM ID alias seems to
be depended on by multiple publishing app charts, and is defined
in Contacts Admin for the sole reason that is is alphabetically
first in the list and is a useful place to inline-define an
alias for the values.

Consequently we've had to move these definitions to the next
publishing app down that uses them, which is Maslow. This is
mildly annoying as Maslow is also marked for retirement and so
we'll have to perform the same dance when we get around to that!

Trello: https://trello.com/c/MiBvG4Xs/3767-retire-contacts-admin