✨ Use makefile to run yamllint command
While working on this repository, it became frustrating to wait for the pipeline to fail before I knew my branch was ready to review. I want fast feedback when working locally, so I created a makefile that conatains a command to install and then run the yamllint command that was continously failing for me. Along with this, I also refactor the ci.yaml file that performs the GHA yamllint for consistency.