✨ Add kubeconform make command
While working in this repository, I found that when pushing changes I
was failing kubeconform requirements. I wanted to test this
locally before pushing, so I've added a make command that lets you
either `make lint` and run all the linting locally, or, run `make
kubeconform` that performs the kubeconform command for you and outputs
the errors locally. This way I was able to quickly identify where the
failure was located and fix it. In this instance, it appeared in the
rendered deployment.yaml file for reports-prototype.