Add `publishing-api` to Plek unprefixable hosts
Plek adds `draft-` to the start of the hostname for Publising API and
the Publishing API Read Replica when running in the draft stack.

There is only one instance of Publishing API for both draft and live, so
we need to specify this application as 'unprefixable' so this prefix is
no longer added.