Temporarily stop db sync for GOV.UK Chat integration env
We're temporarily suspending the db sync for the GOV.UK Chat integration
environment to allow for testing by the rainbow content team.

This will be taking place between 21/07/2025 and 25/07/2025. To ensure the
data doesn't get wiped before we export the data to BigQuery (we have to
run the export manually on int) we're going to suspend the sync for the
duration of the testing.