High5xxRate alert: use fastly_origin_status_group_total
"fastly_origin_status_group_total" is a reasonably new metric which reports the
number of responses from origins, grouped by their status group. Previously we
used "fastly_rt_status_group_total" which was reporting the total number of
responses sent by Fastly, including cached responses.

This was causing situations where the number of 5xx's being reported was a
multiple of the total number of origin requests, because Fastly was serving 3
cached responses for every request that made it to the origin, for example.
This resulted in an alerts reporting error rates in excess of 100%.