Create stripped-down helm chart for publisher-on-pg
As we only need a shell app with database access to postgres, we
shouldn't need all the external connections. However we don't know
which one of these will cause failures, so this is a helm chart
designed to see if we can get away with empty string values in
the globals.