Add ANONYMOUS_USER_ID_SECRET to "publishing" apps in integration
The idea here is to provide this secret to all the apps where users sign
in using signon, so that we can calculate anonymous user ids and send
them to analytics.

I've added the new env var at the top of the list of extraEnv in each
case, because it comes first alphabetically and some of them look like
they're making some rough attempt to be ordered alphabetically.