Add additional sidekiq processes to GOV.UK Chat
We're making changes to GOV.UK Chat to ensure user questions are prioritised
over non-time-sensitive tasks.

We're introducing a bunch of jobs that will run after an answer is composed
for analysis. These need to run on a separate Sidekiq process to ensure
we don't block users from getting their questions answered.

The plan is:

1. merge the PR on GOV.UK Chat that adds the sidekiq config
2. merge this PR
3. merge the PR on GOV.UK Docker that adds the new Sidekiq processes
4. merge the PR on GOV.UK Chat that uses the new Sidekiq processes
5. create a PR to remove the unused worker process here
6. remove unused config on GOV.UK Chat