Run search evaluations one hour later
The evaluations take approximately two hours to run.
This means that the binary evaluation at 8am failed every morning with
an AlreadyExistsError as the evaluation started at 7am was still
running.

Changing all of the evaluations to start one hour later is a temporary
measure to stop them erroring whilst a way to check the status of the
previous job is developed.