Add cron task to send GOV.UK Chat daily activity to Slack
This cronjob will run at 7:13am each morning and call the
`slack:send_previous_days_activity` Rake task in the govuk-chat
application. The 3 mins past the hour is so that we spread jobs out a
bit that all need to run at around the 7am mark.

It's currently commented out as the app still isn't public yet. Once we
go live, we can uncomment this and the other cronjobs too.