Disable GOV.UK Production -> Integration syncing
There is a cross government test running through September and Octover that will use integration as a test environment.

We have been asked to disable the wiping of integration throughout the duration of this test so that the users can update and modify content related to their work without having to recreate it each week, saving time and providing a more realistic experience.

This commit disables the database sync so all publishing apps do not get overwritten. Currently this runs weekly. It is disabled now for the duration of the tests.