Suppress warning notifications when critical alerts are firing
The search team have set up two separate alert thresholds for metrics
related to search quality evaluations results, to distinguish between a
"warning" level drop in quality, and a "critical" level drop in quality.
When the critical alert is firing, it would be overly noisy and possibly
confusing to also notify for the warning alert so this change suppresses
the warning notifications when the critical alert is firing.