Only run binary evaluations every two hours
Reverts: https://github.com/alphagov/govuk-helm-charts/pull/3556

Adds back the binary id so that the rake task will only fetch
evaluations for the binary table every two hours, not all of them.

We have a limit of 20 evaluations per day on production, each table runs
two evaluations, so we have reached our limit. We should contact Google
to increase the evaluation limit to 100 as it is in integration, but it
takes a few days for the requests to be actioned.

See: https://github.com/alphagov/search-api-v2/blob/main/lib/tasks/quality.rake#L23