Add last_over_time recording rule to search quality alerts
Our alerts are flapping off and on because of data dropouts in
Prometheus. Prometheus treats a lack of data as “everything is ok”
by default, so when we’re having an incident and then there’s a
data dropout, the alert stops firing (and is marked as resolved).
When we start getting data again and it’s below the threshold,
the alert becomes active again.

Adding a last_over_time recording rule allows us to smooth over data
gaps by looking at the last metric received over a specified period.
In this case we've choosen 48hrs because some of our metrics are
only updated once a day (so this configuration makes us
resilient to one missed metric).

We've also added a check on absent data, so that when alerts have
been active, they remain active for any length of data dropout.
The check on absent data will also raise a new alert if there is
no data for longer than 48 hours, even if the alert wasn't firing
before the data drop out.