Add last_over_time recording rule to search quality alerts
Our alerts are flapping off and on because of data dropouts in
Prometheus. Prometheus treats a lack of data as “everything is ok”
by default, so when we’re having an incident and then there’s a
data dropout, the alert stops firing (and is marked as resolved).
When we start getting data again and it’s below the threshold,
the alert becomes active again.

Adding a last_over_time recording rule allows us to smooth over data
gaps by looking at the last metric received over a specified period.
In this case we've chosen 26 hours because data drops outs that
are roughly 24 hours in length appear to be quite common for the
clickstream query set.

See Prometheus docs on <aggregation>_over_time() recording rules:
https://prometheus.io/docs/prometheus/latest/querying/functions/#aggregation_over_time