Set readiness probe endpoint for GOV.UK Chat in integration
In Chat, we have a `/healthcheck/ready` endpoint that does various
healthchecks that want to be performed continuously.

Currently Chat has all 3 healthcheck endpoints configured to
`/healthcheck/live` as this is the default. This means our
`/healthcheck/ready` endpoint isn't doing anything.

This sets the endpoint for integration only, so that we can do a test
first.