Remove superfluous labels on search quality alert rules
When adding the last_over_time recording rule to the
search quality alerts, it looked as though the
labels relating to the evaluation metrics ("dataset",
"month" and "top") had to be manually readded as part of
the alert rule. In fact they don't, and the "top" label is
causing an error in the deploy because it's not a string.
Therefore, we're removing the superfluous configuration
of these labels in the alert rule.