Bump frontend replica count
Frontend has been accruing formats recently, but we haven't applied any more resources to support the extra work.

We're going to do some better capacity management planning to better understand the shape of the different resources required (memory, cpu, connections, etc), but for now, we just want to add some extra nodes.

This shouldn't add any particular load to downstream services, such as content-store, as the demand is the same, it's just shifted.