Set `DB_OWNER` for Content Block Manager restore
The database owner of Content Block Manager is `content-block-manager`
(with dashes), but the `default_db_owner` function assumes the owner’s
username is the name of the database with the dashes replaced with
underscores (in this case `content_block_manager`). To get around this,
we add an extra env variable to ensure we override the `DB_OWNER`
variable, which should solve our problem.