Set `DB_OWNER` in correct environment
In https://github.com/alphagov/govuk-helm-charts/pull/3601 I set the `DB_OWNER`
env variable in an attempt to get the database restore jobs working.
However, I put this in only the production environment, where the
restore job doesn’t run.

This removes the env var from production and moves it to staging and
integration, where it will be used.