Add govuk-e2e-test jobs to staging and production
so they can be run manually.  Otherwise they don't appear on the list of
cronjobs (kubectl get cronjobs --all-namespaces).  Attempt of running returns
an error:

  kubectl create job govuk-e2e-tests-manual-run --from=cronjob/govuk-e2e-tests
  Error from server (NotFound): cronjobs.batch "govuk-e2e-tests" not found

The value of `schedule` also needs to be a valid cron expression so I'm setting
them to what they used to be before they got removed in
https://github.com/alphagov/govuk-helm-charts/commit/9dab0b5fabc7c14ec1fabc64f1317ab2c65dc576