Alert only on the longest validity Signon token
This alert tells us if any tokens expire in less than 14 days in Signon,
based on the nearest expiring token.

We should work on the assumption that we are always using the token with
the longest validity, so don't need to alert on older tokens expiring as
they have already been rotated.

This means we will not get alerts when a token has automatically been
rotated by the script in https://github.com/alphagov/signon/pull/4213.