Add nginx config to GOV.UK Chat endpoint to turn off buffering
This is a test to see whether this config works for nginx to allow it to
buffer a response.

Currently - and by default - nginx will buffer all responses, i.e. it
waits until it's got the entire response from downstream before sending
it to the client.

We're spiking out a feature in GOV.UK Chat where responses are streamed
to the client (like most chatbots do). The nginx config is preventing
that from happening.

This commit adds what I think should be enough config to turn off
buffering. It's currently just applied to one endpoint in GOV.UK Chat,
and it's only set in integration.