Add support for websockets in nginx config for Chat
After some investigation, it seems that the nginx config for Chat needs to
be modified to support websockets properly. The changes involve adding a new
location block for /cable that correctly proxies websocket connections to the
Chat service.

I've found a couple of examples which have led my approach:

https://gist.github.com/SunDi3yansyah/5978a1737d08f4abda46f1176a039cc2
https://stackoverflow.com/questions/67696655/what-do-i-need-to-do-to-hook-up-actioncable-on-nginx-and-puma

One thing i noticed is that when i do "kubectl describe svc govuk-chat"
the port is listed as 8080, not 3000 as i had assumed. So I've updated the
configuration to reflect that.