Set additional proxy headers for GOV.UK Chat streaming
These headers have been copied from the top-level `location` directive
in the nginx config.

Currently browsing to this path - i.e. the path ending `/answer_stream`
gives a 403 because of a disallowed host. We need to set the host in the
header to be the original host of the request.

I've also copied the other directives from the top-level `location /`
directive just in case those are needed too.