Only show this_month on search quality dashboards
Previously the dashboard panels for search quality
metrics Binary recall@3 and Clickstream NDCG@10 showed
data for this_month and last_month, since each day
evaluations are run against this month's and last
month's Sample Query Sets. However, the alerting that
notifies us on drops in these metrics only refers to
this month's data. This change adds an additional
filter so that the dashboard panels only show data
for this month, so that they are more clear for someone
responding to an alert.