Reimport graphql-stats dashboard from Grafana prod
I'm not sure why, but the dashboard in production seems to have diverged
a bit from what's in the infra-as-code. I'd like to make some changes to
the dashboard, and I don't want the diff to include all this divergance,
so here's a commit just to bring them in line.