Add app management scripts for maintenance
This includes a set of python scripts that manipulate the values file for a given environment.

Each script takes a list of apps to perform a given operation on, and the path to a values file.

Each operation either enables or disables app deployments, worker deployments or cron jobs.

This is done via string manipulation of the yaml file, instead of parsing it and manipulating it as a data structure. This is to preserve formatting in the file, as well as various things like comments and yaml anchors.

These scripts are for use during app maintenance. They are not intended to be used long-term, and as such may be flaky or difficult to maintain.