Increase govuk-e2e-test memory in staging and production
The manual job is triggered to validate the system is working following a
restart of the apps. The apps are slower after restarting (no cache) and
consequently tests take longer to execute resulting in higher memory usage.

This should prevent pods from exceed its memory limit, which lead to OOMKilled.