Reformat values-production.yaml to work around yq bug
Makes minor changes to certain multiline strings to work around a bug in yq

In the future, yq will be used to manipulate this file programatically

yq seems to have a bug with multiline > strings where it inserts a newline in the middle of the string each time yq writes out the file. This causes a large amount of newlines to appear in the file after multiple yq operations.