Replace app management scripts with a single manage-yq.py script
This new script performs the function of all of the removed scripts

It uses yq instead of string manipulation, which drastically simplifies the script

This new script takes a CSV export of the app categories spreadsheet.
This allows the user to specify which components of each app should be
enabled/disabled.