Update GOV.UK Chat Worker autoscaling to target answer queue
At the moment, the autoscaling for the GOV.UK Chat Worker is targeting
the queue length for all jobs. We're adding a bunch of jobs which will
handle auto-evaluation. These will be processed by the default-worker
process. We don't want the autoscaling to respond to the backlog
of these auto-evaluation jobs as they're not time senseitive.

This updates the MetricsQuery and SeriesQuery in the Prometheus Adapter
values files to only target the answer queue.