Update Slack alerts for Chat AI worker Sidekiq queues
We recently scoped autoscaling to the answer Sidekiq queue. This commit
updates the existing Slack alerts to be scoped to the answer queue.

We don't want to alert on high queue sizes for the default queue since
those jobs aren't time sensitive.

We only care about a delay in the answer queue since those jobs are directly
answering user questions.

I've added a new alert for the default queue if there is a job that is older
than six hours.