Vertex response time alerting for search api v2
Search API v2 sends metrics to Prometheus to allow us to monitor how long it takes
vertex to respond to search requests [1]

This commit adds a recording rule and an alert which is triggered if the 90th percentile of
requests exceeds 2seconds.

The test for the recording rule defines three buckets and the sequence of requests allocated to each to confirm that the rule returns
the value of the 90th percentile. There is documentation on historgram_quantiles at [2], which I only understand in a very minimal way.

[1] https://github.com/alphagov/search-api-v2/blob/7dc0b5fc8597b9827482e71127b4fcc4871e86d9/app/services/metrics/exported.rb#L24-L26
[2] https://prometheus.io/docs/prometheus/latest/querying/functions/#histogram_quantile