Update environment variables to send no traffic to GraphQL
We are upgrading the Postgres database for Publishing API. This is
likely to have a few minutes downtime.

Therefore we will switch off all traffic to the GraphQL API prior to
this, to remove the risk of errors being served to public users.

This will be reverted once the upgrade is complete.